requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.6
cryptography==45.0.7
Django==5.2.6
django-cors-headers==4.8.0
django-oauth-toolkit==3.0.1
djangorestframework==3.16.1
idna==3.10
jwcrypto==1.5.6
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
